Correct missed rename
