+++ pyccel/codegen/printing/ccode.py
@@ -33,9 +33,9 @@ from pyccel.ast.literals  import LiteralString, LiteralInteger, Literal, Nil
 from pyccel.ast.numpyext import NumpyFull, NumpyArray
 from pyccel.ast.numpyext import NumpyReal, NumpyImag, NumpyFloat
 
-from pyccel.ast.variable  import IndexedVariable, Slice, ValuedVariable
-from pyccel.ast.variables import PyccelArraySize, Variable, VariableAddress
-from pyccel.ast.variables import DottedName
+from pyccel.ast.variable import IndexedVariable, Slice, ValuedVariable
+from pyccel.ast.variable import PyccelArraySize, Variable, VariableAddress
+from pyccel.ast.variable import DottedName
 
 
 from pyccel.codegen.printing.codeprinter import CodePrinter
